Added a new feature to link the message board feature to a mailing list service.
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@19112 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portlet/calendar/service/impl/CalEventLocalServiceImpl.java
@@ -1222,7 +1222,7 @@ public class CalEventLocalServiceImpl extends CalEventLocalServiceBaseImpl {
 
 		int recurrenceType = recurrence.getFrequency();
 
-		int interval = recurrence.getInterval();
+		int interval = (int)recurrence.getInterval();
 
 		if (recurrenceType == Recurrence.DAILY) {
 			recur = new Recur(Recur.DAILY, -1);
@@ -1450,7 +1450,7 @@ public class CalEventLocalServiceImpl extends CalEventLocalServiceBaseImpl {
 				addField = Calendar.YEAR;
 			}
 
-			int addAmount = recurrence.getInterval() * recur.getCount();
+			int addAmount = (int)recurrence.getInterval() * recur.getCount();
 
 			until.add(addField, addAmount);
 			until.add(Calendar.DAY_OF_YEAR, -1);
